Fixed test for multiple apps
